libpod: export pod rendering and creation to a lib
