+++ tests/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
         "job.go",
         "login.go",
         "ping.go",
-        "pod_servers.go",
         "test.go",
         "utils.go",
         "vmi_servers.go",
@@ -34,6 +33,7 @@ go_library(
         "//tests/containerdisk:go_default_library",
         "//tests/flags:go_default_library",
         "//tests/libnet:go_default_library",
+        "//tests/libpod:go_default_library",
         "//tests/libvmi:go_default_library",
         "//tools/vms-generator/utils:go_default_library",
         "//vendor/github.com/Masterminds/semver:go_default_library",
@@ -199,6 +199,7 @@ go_test(
         "//tests/containerdisk:go_default_library",
         "//tests/flags:go_default_library",
         "//tests/libnet:go_default_library",
+        "//tests/libpod:go_default_library",
         "//tests/libvmi:go_default_library",
         "//tests/network:go_default_library",
         "//tests/reporter:go_default_library",
